refactor: refactor for better readability
